tests: move broker related stuff out of neon_fixtures.py
